requirements.txt
openai>=1.0.0
librosa>=0.10.0
numpy>=1.21.0
pathlib2>=2.3.0
requests>=2.28.0
mutagen>=1.45.0
vaderSentiment>=3.3.2
Pillow>=9.0.0 